backend/requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database & ORM
supabase==2.9.0

# AI & LLM
openai==1.54.5
langchain==0.3.7
langchain-openai==0.2.9

# Utilities
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# CORS
python-multipart==0.0.17

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Testing (additional)
pytest-mock==3.14.0
faker==30.8.2
